# fix: SQL injection prevention performance optimization - resolves Q1-…